+++ test.rb
@@ -1,7 +1,0 @@
-def foo
-  begin
-    a
-  rescue
-    b
-  end
-end